feat: added productdetail isi data
